from double precision to single precision

+++ private_local_memory.cl
@@ -1,15 +1,15 @@
 //#pragma OPENCL EXTENSION cl_khr_fp64: enable
 __kernel //__attribute__ ((reqd_work_group_size(2048, 2048, 1)))
-void private_local_memory(__global double* a, __global double* b, __global double* output, __local double* B_local)
+void private_local_memory(__global float* a, __global float* b, __global float* output, __local float* B_local)
 {
     int r = get_global_id(0);
     int c, k, index;
-    double running;
+    float running;
     int rank = get_global_size(0);
     int iloc = get_local_id(0); 
     int nloc = get_local_size(0);
 
-    double A_private[4096];
+    float A_private[4096];
 
     for(index = 0; index < rank; index++){
         A_private[index] = a[r*rank + index];
